% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtriangle.R
\name{find.area.rtriangle}
\alias{find.area.rtriangle}
\title{Title}
\usage{
\method{find.area}{rtriangle}(atringle)
}
\arguments{
\item{atringle}{}
}
\value{
numeric: the area of the triangle
}
\description{
Title
}
\examples{
atriangle <- rtriangle(3,4); find.area(atriangle)
}

