% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtriangle.R
\name{find.hypotenuse.rtriangle}
\alias{find.hypotenuse.rtriangle}
\title{Title}
\usage{
\method{find.hypotenuse}{rtriangle}(atriangle)
}
\arguments{
\item{atriangle}{}
}
\value{
numeric: perimeter of triangle
}
\description{
Title
}
\examples{
atriangle <- rtriangle(3,4); find.hypotenuse(atriangle)
}

