% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtriangle.R
\name{find.perimeter.rtriangle}
\alias{find.perimeter.rtriangle}
\title{Title}
\usage{
\method{find.perimeter}{rtriangle}(atriangle)
}
\arguments{
\item{atriangle}{}
}
\value{
numeric:
}
\description{
Title
}
\examples{
atriangle <- rtriangle(3,4); find.perimeter(atriangle)
}

