% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{plot.pslg}
\alias{plot.pslg}
\title{Plot pslg object}
\usage{
\method{plot}{pslg}(x, ...)
}
\arguments{
\item{x}{\code{\link{pslg}} object}

\item{...}{Arguments to be passed to methods.}
}
\description{
Plot \code{\link{pslg}} object
}
\author{
David Sterratt
}
