\name{plot.triangulation}
\alias{plot.triangulation}
\title{Plot a triangulation object produced with triangulate}
\usage{\method{plot}{triangulation} (x, ...)}
\description{Plots a triangulation object produced with \code{\link{triangulate}}}
\author{David Sterratt}
\arguments{\item{x}{Triangulation object produced with \code{\link{triangulate}}.}
\item{...}{Arguments to be passed to methods.}}
