\name{read.pslg}
\alias{read.pslg}
\title{Read a Planar Straight Line Graph from file}
\usage{
read.pslg(file)
}
\arguments{
  \item{file}{File name of \code{.poly} file to read.}
}
\value{
\code{pslg} object. See \code{\link{pslg}}.
}
\description{
Read a Planar Straight Line Graph from a \code{.poly} file,
as used in Shewchuk's Triangle library
(\url{http://www.cs.cmu.edu/~quake/triangle.poly.html}).
}
\author{
David Sterratt
}

